+++ webpack.base.config.js
@@ -1,3 +1,8 @@
+/**
+ * Copyright 2016 aixigo AG
+ * Released under the MIT license.
+ * http://laxarjs.org/license
+ */
 /* eslint-env node */
 
 const path = require( 'path' );