+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Last Changes
 
+
+## v2.0.0-alpha.2
+
 - [#311](https://github.com/LaxarJS/laxar/issues/311): moved webpack-jasmine-html-runner-plugin to its own project
 - [#312](https://github.com/LaxarJS/laxar/issues/312): provide required polyfills as bundle
     + NEW FEATURE: see ticket for details